HTCONDOR-1925 fix Fedora warnings
There's a new warning in the fedora compiler where operator new[]
takes a size_t as the size, and if passed a (signed) int, it
gets promoted to a unsigned quantity, which if it was negative,
(unlikely), gets changes to a very large integer, which triggers
a warning

+++ src/condor_schedd.V6/qmgmt.cpp
@@ -3097,7 +3097,7 @@ grow_prio_recs( int newsize )
 
 	dprintf(D_FULLDEBUG,"Dynamically growing PrioRec to %d\n",newsize);
 
-	tmp = new prio_rec[newsize];
+	tmp = new prio_rec[(unsigned int)newsize];
 	if ( tmp == NULL ) {
 		EXCEPT( "grow_prio_recs: out of memory" );
 	}
